back/requirements.txt
Django
python-dotenv
djangorestframework
django-cors-headers
pymysql